# docs/openapi/openapi.yaml
basePath: /
definitions:
...
  connector.PromptInput:
    properties:
      agentConfig:
        additionalProperties: {}
        description: AgentConfig is the agent config (agents.config jsonb).
        type: object
      agentID:
        type: string
      agentName:
        type: string
      agentSlug:
        type: string
      conversationID:
        type: string
      conversationInitiatorID:
        type: string
      runID:
        description: |-
          RunID is the Parsar agent_run row id; the connector uses it
          to scope any per-run scratch state.
        type: string
      triggerAttachments:
        description: |-
          TriggerAttachments carries non-text payloads (today: images
          downloaded from a Feishu inbound message) that arrived alongside
          TriggerMessageContent. Connectors that don't understand multimodal
          input may safely ignore the field.
        items:
          $ref: '#/definitions/store.MessageAttachment'
        type: array
      triggerMessageContent:
        description: |-
          TriggerMessageContent is the user-facing message that drives this
          prompt. The connector decides how to fold it into its own prompt
          schema.
        type: string
      workspaceID:
        type: string
    type: object
...
  store.MessageAttachment:
    properties:
      data_base64:
        description: |-
          DataBase64 is the standard-base64 encoded raw bytes. Callers
          MUST NOT base64-encode again on persistence.
        type: string
      kind:
        type: string
      mime:
        description: |-
          MIME is the upstream-reported content type. Falls back to a
          kind-specific default when upstream omits it.
        type: string
      size:
        description: |-
          Size is the byte count before base64 expansion. Cached so
          observability surfaces don't have to decode.
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    Parsar harness API. Optional Feishu SSO and event delivery
    are enabled only when the corresponding PARSAR_FEISHU_* values are configured.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Parsar API
  version: 0.1.0
paths:
...
  /dev/connectors/opencode/stream:
    post:
      consumes:
      - application/json
      description: Dev-only Server-Sent Events endpoint that forwards the OpenCode
        connector's StreamPrompt output. Does NOT manage agent_run state — callers
        drive lifecycle via the regular dev endpoints.
      operationId: streamDevConnectorPrompt
      parameters:
      - description: Prompt input (workspace_id, conversation_id, run_id, agent_config
          required)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/connector.PromptInput'
      produces:
      - application/x-json-stream
      responses:
        "200":
          description: SSE stream opened
        "400":
          description: Body decode or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: ResponseWriter does not support flushing or connector failure
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Connector does not support streaming
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: OpenCode AgentConnector is not registered
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream a raw connector prompt (dev SSE)
      tags:
      - gateway
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    in: cookie
    name: parsar_session
    type: apiKey
swagger: "2.0"